Fix unreadable images uploaded in forum
The urls of images uploaded to the course forums are stripped from
"?...&..." parameters (see django_comment_client.base.views.upload).
Thus, the only way to access those images is to make sure that they are
uploaded to a public bucket. It turns out the default open edx bucket
must be public (but not the ora2 bucket).