feat: upgrade to maple
The main change here is that by upgrading MinIO to the latest release, the
console now runs as a different web service. Thus, we need to expose the MinIO
file storage server as a different host. We decided to switch to files.LMS_HOST
because it is more consistent with how other apps use S3.