fix: user task storage (course export, reports)
edx-platform course task storage was defaulting to non-authenticated urls. This
caused us to go through a bug in django-storages where querystring arguments
were not properly escaped. In any case, it's a (much) better idea to return
authenticated urls with expiry. We could make use of ImportExportS3Storage, but
that class makes use of the boto library, which does not support sig3v4
authentication. Thus, we need to create a custom class which will emulate the
behaviour of that class, but will use boto3 instead.

See: https://discuss.overhang.io/t/export-library-and-course-error-with-minio-s3-integrated/2338
https://discuss.overhang.io/t/issue-with-minio-with-overhang/2362

Close #13.