feat: restore Azure gateway compatibility
MinIO dropped gateway compatibility in May 2022:
https://blog.min.io/deprecation-of-the-minio-gateway/
https://github.com/minio/minio/releases/tag/RELEASE.2022-04-30T22-23-53Z

To restore compatibility, we downgrade the minio and mc images to
pre-deprecation. This is a rollback of just two versions, so existing
users should be fine.

However, defining bucket policy fails in gateway mode with Azure. Thus,
we bypass the `mc policy` init command when running as a gateway to
Azure. We add a warning to the docs to tell Azure users to modify this
policy by hand.