fix: make LMS/Studio connect to MinIO on port 9000 in dev
The value for AWS_S3_ENDPOINT_URL set in openedx-development-settings
specifies port 9000, and it was being written to Studio and LMS settings
correctly. But this is the sequence of what was really happening:

1. openedx-common-settings first sets AWS_S3_ENDPOINT_URL WITHOUT port
   info, since this is how it's used in prod deployments.
2. openedx-common-settings imports S3Boto3Storage because it wants to
   make a subclass to use for USER_TASKS_ARTIFACT_STORAGE.
3. The act of importing that pacakge has the side-effect of initializing
   Django storages and sets up the default storage object
   (django.core.files.storage.default_storage).
4. Because the DEFAULT_FILE_STORAGE class is S3Boto3Storage, it then
   reads the value of AWS_S3_ENDPOINT_URL (which has no port info), and
   sets this as the S3 connection information for the S3Boto3Storage
   instance at that time.
5. openedx-development-settings then comes along and changes the value
   of AWS_S3_ENDPOINT_URL to specify port 9000, but by this point it's
   too late to affect django-storages (though it will confuse folks who
   try to debug). The value is ignored.

This commit removes the import that was causing the premature Django
storages initialization. The patch is no longer necessary because
ImportExportS3Storage was already moved to use S3Boto3Storage in
edx-platform 31002ab (it is included in the Palm release).